Re-enable status bar updating
This approach is still timer based as a 1 second update time
seems pretty reasonable. Events would be ideal, but this would
probably make the update speed significantly faster an make
using the the visual cue for video synchronization less
effective. This can be revisited at a later date if necessary.